+++ webserver/web-interface/src/components/profile/edit/ProfileEdit.tsx
@@ -14,11 +14,11 @@ import GlobeIcon from '@mui/icons-material/Public';
 import PreviousIcon from '@mui/icons-material/NavigateBefore';
 import {
   CurveStyle, Phase, PhaseType, Profile,
-} from '../../models/profile';
-import ProfileChart from '../chart/ProfileChart';
-import { GlobalRestrictions } from './GlobalRestrictions';
+} from '../../../models/profile';
+import ProfileChart from '../../chart/ProfileChart';
+import { GlobalRestrictions } from '../GlobalRestrictions';
 import PhaseEdit from './PhaseEdit';
-import { getIndexInRange } from '../../models/utils';
+import { getIndexInRange } from '../../../models/utils';
 import { ProfileGlobalsEdit } from './ProfileGlobalsEdit';
 
 export interface ProfileEditProps {
