+++ webserver/web-interface/src/components/profile/edit/ProfileGlobalsEdit.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Typography } from '@mui/material';
 import Grid from '@mui/material/Unstable_Grid2/Grid2';
-import { Profile, GlobalStopConditions } from '../../models/profile';
-import { SettingsNumberInput } from '../Tabs/settings_inputs';
-import { constrain } from '../../models/utils';
+import { Profile, GlobalStopConditions } from '../../../models/profile';
+import { SettingsNumberInput } from '../../inputs/settings_inputs';
+import { constrain } from '../../../models/utils';
 
 export interface ProfileGlobalsEditProps {
   profile: Profile;
